Create directory for python tests
