+++ solutions/python/largest-continuous-sum.py
@@ -1,4 +1,3 @@
-import unittest
 """solution to the largest-continuous-sum problem"""
 
 def largest_continuous_sum(arr):
@@ -14,17 +13,3 @@ def largest_continuous_sum(arr):
             largest = sum
         last = num
     return largest
-
-class LargestContinousSequenceTest(unittest.TestCase):
-    def test_largest_continuous_sum(self):
-        sum = largest_continuous_sum([1, 2, 3, 4])
-        self.assertEqual(sum, 10)
-
-        sum = largest_continuous_sum([1, 2, 3, 1, 2, 3, 4])
-        self.assertEqual(sum, 10)
-
-        sum = largest_continuous_sum([1, 2, 3, 2, 5, 4])
-        self.assertEqual(sum, 6)
-
-if __name__ == '__main__':
-    unittest.main()
